// src/types/SocialLinks.ts
// ✅ src/types/SocialLinks.ts

export interface SocialLinks {
    facebook?: string;
    instagram?: string;
    linkedin?: string;
    youtube?: string;
    xing?: string;
    twitter?: string;
    tiktok?: string;
    pinterest?: string;
    telegram?: string;
    whatsapp?: string;
    [key: string]: string | undefined;
  }
  